requirements.txt
medmnist
PIL
numpy
matplotlib.pyplot
tensorflow